Update tablearray.csv
The entry for "You'll Shoot Your Eye Out!" had a curly apostrophe in the name rather than a straight apostrophe. All other table names that have an apostrophe have a straight apostrophe. The table "You'll Shoot Your Eye Out!" wasn't showing up in the VPX Launcher list because VPX Launcher was interpreting the apostrophe in the name as a question mark character and wasn't matching. Therefore, the only way to display the table in the list was under the Filebrowser filter. By changing the curly apostrophe to a straight apostrophe, I'm hoping that the table will then be listed with the rest of the tables.